# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "Island Go API",
        "contact": {},
        "license": {},
        "version": "1.0"
    },
    "host": "127.0.0.1:3000",
    "paths": {
        "/api/v1/article": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "文章"
                ],
                "summary": "新建文章",
                "parameters": [
                    {
                        "description": "文章",
                        "name": "article",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/service.ArticleCreateService"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/serializer.ArticleResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/article/{id}": {
            "get": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "文章"
                ],
                "summary": "文章详情",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/serializer.ArticleResponse"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "文章"
                ],
                "summary": "更新文章",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "更新",
                        "name": "update",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/service.ArticleUpdateService"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/serializer.ArticleResponse"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "文章"
                ],
                "summary": "删除文章",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/serializer.Response"
                        }
                    }
                }
            }
        },
        "/api/v1/articles": {
            "get": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "文章"
                ],
                "summary": "文章列表",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "每页数量",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "当前页数",
                        "name": "page",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/serializer.ArticleListResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "serializer.Article": {
            "type": "object",
            "required": [
                "content",
                "profile",
                "tags",
                "title"
            ],
            "properties": {
                "content": {
                    "description": "内容",
                    "type": "string"
                },
                "createdAt": {
                    "description": "创建时间",
                    "type": "integer"
                },
                "creator": {
                    "description": "作者",
                    "type": "object",
                    "$ref": "#/definitions/serializer.User"
                },
                "image": {
                    "description": "banner图",
                    "type": "string"
                },
                "profile": {
                    "description": "简介",
                    "type": "string"
                },
                "tags": {
                    "description": "标签",
                    "type": "string"
                },
                "title": {
                    "description": "标题",
                    "type": "string"
                },
                "topic": {
                    "description": "话题",
                    "type": "object",
                    "$ref": "#/definitions/serializer.Topic"
                }
            }
        },
        "serializer.ArticleItem": {
            "type": "object",
            "required": [
                "profile",
                "tags",
                "title"
            ],
            "properties": {
                "createdAt": {
                    "description": "创建时间",
                    "type": "integer"
                },
                "creatorAvatar": {
                    "description": "作者头像",
                    "type": "string"
                },
                "creatorName": {
                    "description": "作者名字",
                    "type": "string"
                },
                "image": {
                    "description": "banner图",
                    "type": "string"
                },
                "profile": {
                    "description": "简介",
                    "type": "string"
                },
                "tags": {
                    "description": "标签",
                    "type": "string"
                },
                "title": {
                    "description": "标题",
                    "type": "string"
                },
                "topicName": {
                    "description": "话题名称",
                    "type": "string"
                }
            }
        },
        "serializer.ArticleListResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "description": "状态码",
                    "type": "integer",
                    "example": 2000
                },
                "count": {
                    "type": "integer"
                },
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/serializer.ArticleItem"
                    }
                },
                "message": {
                    "description": "自定义消息",
                    "type": "string",
                    "example": "success"
                }
            }
        },
        "serializer.ArticleResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "description": "状态码",
                    "type": "integer",
                    "example": 2000
                },
                "data": {
                    "type": "object",
                    "$ref": "#/definitions/serializer.Article"
                },
                "message": {
                    "description": "自定义消息",
                    "type": "string",
                    "example": "success"
                }
            }
        },
        "serializer.Response": {
            "type": "object",
            "properties": {
                "code": {
                    "description": "状态码",
                    "type": "integer",
                    "example": 2000
                },
                "data": {
                    "description": "响应数据",
                    "type": "object"
                },
                "message": {
                    "description": "自定义消息",
                    "type": "string",
                    "example": "success"
                }
            }
        },
...
        "serializer.Topic": {
            "type": "object",
            "properties": {
                "icon": {
                    "description": "图标",
                    "type": "string"
                },
                "image": {
                    "description": "背景",
                    "type": "string"
                },
                "info": {
                    "description": "详情",
                    "type": "string"
                },
                "title": {
                    "description": "标题",
                    "type": "string"
                }
            }
        },
...
        "serializer.User": {
            "type": "object",
            "properties": {
                "avatar": {
                    "description": "用户头像",
                    "type": "string"
                },
                "createdAt": {
                    "description": "创建时间",
                    "type": "integer"
                },
                "email": {
                    "description": "用户邮箱",
                    "type": "string"
                },
                "id": {
                    "description": "用户ID",
                    "type": "integer"
                },
                "status": {
                    "description": "用户状态",
                    "type": "integer"
                },
                "username": {
                    "description": "用户名称",
                    "type": "string"
                }
            }
        },
...
        "service.ArticleCreateService": {
            "type": "object",
            "required": [
                "content",
                "profile",
                "tags",
                "title",
                "topicId"
            ],
            "properties": {
                "content": {
                    "description": "内容",
                    "type": "string"
                },
                "image": {
                    "description": "banner图",
                    "type": "string"
                },
                "profile": {
                    "description": "简介",
                    "type": "string"
                },
                "tags": {
                    "description": "标签",
                    "type": "string"
                },
                "title": {
                    "description": "标题",
                    "type": "string"
                },
                "topicId": {
                    "description": "话题ID",
                    "type": "integer"
                }
            }
        },
        "service.ArticleUpdateService": {
            "type": "object",
            "required": [
                "content",
                "profile",
                "tags",
                "title",
                "topicId"
            ],
            "properties": {
                "content": {
                    "description": "内容",
                    "type": "string"
                },
                "image": {
                    "description": "banner图",
                    "type": "string"
                },
                "profile": {
                    "description": "简介",
                    "type": "string"
                },
                "tags": {
                    "description": "标签",
                    "type": "string"
                },
                "title": {
                    "description": "标题",
                    "type": "string"
                },
                "topicId": {
                    "description": "话题ID",
                    "type": "integer"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "token",
            "in": "header"
        }
    }
}